CRT-829 slight refactor
